How this works: - store active chats (updated in recent <24h) in chatData.chats as normal. - maintain an array chatData.chatNames that stores all chat names, regardless of active/inactive. - for active chats, same operations as before; - for inactive chats, when using them, we first retrieve it from localstorage, then mark them as active (ie. add chat to chatData.chats). don't forget to clear up localstorage.
Note how a chat will switch from inactive to active when the user accesses it.

Note how the extension checks the following when a message is set:
compares the current time with lastCleanupTime; if at least a fixed duration (eg. 6h) has passed then:
we iterate over all active chats and checks if they may be inactive.

**NOT WORKING.**