Add Google Gemini provider back (#28)
The removal of Gemini was probably not the best idea... >:)
1. Google API Keys are completely free and the rate limits are pretty
generous: https://ai.google.dev/gemini-api/docs/models/gemini
2. Gemini models (specifically gemini-1.5-flash) are pretty good

### huge update: addition of streaming
Part of the reason why Gemini was removed is because I didn't write
streaming support for it, and that made it pretty subpar. But now I have
also added **streaming support** for it!! To make this possible, I
refactored a sizable chunk of code. There is some added complexity, but
hopefully it's organised well enough. I've also tried my best to ensure
that nothing should break when it comes to backward compatibility.

We now also support multiple Gemini API Keys in the preferences, so the
rate limit is effectively increased.