Switch to gemini-g4f package (#36)
This package is maintained by myself. Currently it's pretty much the
same as gemini-ai, but I'll be adding some custom fixes and improvements
in the future. I will also constantly update it when gemini-ai receives
updates.

Right now, there's a very minor speedup added - since in this extension
we use node-fetch which supports AsyncIterators, we default to using
those in gemini-g4f.