Add support for uploading files (#44)
Added support for uploading files. This includes AI Chat, several AI
commands, as well as optionally in Custom AI Commands.

To upload a file in AI Chat, press Command-Enter or select "Compose
Message" from the actions. Then, simply click on the upload button to
get started.
To upload a file in an AI command, first ensure you're not selecting any
text, and then simply run the AI command and click on the upload button.

**Currently, this is limited to the Google Gemini Provider only, but I
expect to add more providers in the future as the trend shifts towards
multimodality.**

More changes:
1. Now there is a dedicated providers module that unifies all available
providers.
2. Rewrite Gemini provider so as to support files.