Fix double response generation in AI commands (#51)
For unknown reasons, the useEffect hook in gpt.jsx is rendered twice,
both at the same time.
This results in getResponse() being called twice whenever a command is
called directly (i.e. the form is not shown).
This problem has always been present, and it's also present in the
original raycast-gemini.

The problem is reproducible by launching any AI command **with all
arguments passed directly from the root search**. The displayed text
will switch back and forth between two different generating responses.
The problem is especially noticeable when using a provider that streams
chunks quickly.

We work around this by preventing any generation when the
generationStatus.loading flag is set to true.