Remove g4f-ts package, replace with g4f-image (#64)
Remove the `g4f` package completely, as a continuation to #63.

g4f-image contains the image generation features from g4f-ts. It's
maintained by myself, and it's much, much more lightweight than its
predecessor.

This patch also means that all the packages we use for response
generation are now self-maintained, a significant milestone for this
project.