Support rendering as plain text (#136)
In certain contexts, we want to render as plain text - e.g. the Format
Text command.

The Detail view only supports markdown currently, so we perform
modifications on the markdown to render it as plain text. It's not a
perfect implementation, but most of the time it works correctly.

This feature has been added to:
- a new "View as Plain Text" command in AI Chat
- Format Text command