Fix memory usage in AI Chat (#141)
This is a major bug fix.

Previously, out-of-memory issues would sometimes be encountered during
response generation; this occurred especially for longer responses.

This issue was caused by the lack of throttling in the updates to the
chat data; the frequent renders apparently prevented the garbage
collector from functioning properly.

This patch throttles the updates at 100ms intervals, both in AI Chat and
in commands. We use lodash.throttle, a lightweight and functional
library. The memory usage issue has since been fixed, as verified
through comparisons of memory usage logs.

This patch also fixes a bug where Web Search would be called twice in AI
Chat.

Additionally, a Developer Mode option has been added to the preferences.
When enabled, it automatically logs memory usage.