Proxy support (experimental) (#153)
Support using a HTTP(s) proxy. Feature is experimental, and may not be
working for all users currently.

Technical side: Defined a custom `fetch` function that completely
replaces the previous one. All requests now go through this (with the
exception of curl), which allows us to apply the HTTP agent.