Optimise response formatting (#158)
formatResponse is no longer called in the generation function itself.
Instead, it is used by the provider. It makes the generation code less
messy.

This also makes the responses consistent between what the provider
generates, and what is actually displayed.

Lastly, this is a large performance optimisation since formatResponse is
now called only when necessary, instead of being called after every
chunk.