Throttle storage writes in AI Chat (#166)
We don't need to write to storage after every chunk is streamed. We only
update every 30s, and force the update every time the response
completes/stops, a message is sent, or we are switching to a new chat.

Note: I attempted to also update the storage upon exit. But as per a
comment in the slack, "You can't. It's not Node.js. Raycast transpiles
the code to SwiftUI."

So this introduces a **breaking change**: Exiting the extension before
the response has finished will lead to the partial response being lost.