Refactor response generation (#200)
All core functionality of response generation has been moved to
`response.js`.

One new change, in particular, is the `generateResponse` function which
combines all 3 types of providers into a single AsyncGenerator.

This makes the code much more maintainable.

Performance wise, there should be no significant impact, but this needs
to be observed for some time.

As a result of this patch, #193 is also fixed.

fixes #193